Move FakeFiber to separate file
