+++ modules/core/src/main/scala/fs2/kafka/KafkaConsumer.scala
@@ -76,31 +76,6 @@ sealed abstract class KafkaConsumer[F[_], K, V]
     with KafkaMetrics[F]
     with KafkaConsumerLifecycle[F]
 
-case class FakeFiber[F[_]](join: F[Unit], cancel: F[Unit])(implicit F: Concurrent[F]) {
-  def combine(that: FakeFiber[F]): FakeFiber[F] = {
-
-    val fa0join =
-      this.join.guaranteeCase {
-        case Outcome.Canceled() => F.unit
-        case _                  => that.cancel
-      }
-
-    val fb0join =
-      that.join.guaranteeCase {
-        case Outcome.Canceled() => F.unit
-        case _                  => this.cancel
-      }
-
-    FakeFiber(
-      F.racePair(fa0join, fb0join).flatMap {
-        case Left((a, fiberB))  => F.map2(a.embedNever, fiberB.joinWithNever)((_, _) => ())
-        case Right((fiberA, b)) => F.map2(fiberA.joinWithNever, b.embedNever)((_, _) => ())
-      },
-      F.map2(this.cancel, that.cancel)((_, _) => ())
-    )
-  }
-}
-
 object KafkaConsumer {
   private def spawnRepeating[F[_]: Concurrent, A](fa: F[A]): Resource[F, FakeFiber[F]] =
     Resource.make {
